subscription/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.16.2
pydantic[email]==2.9.2
stripe==11.1.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
